Mostly compiles

+++ zebROS_ws/src/gpu_apriltag/include/gpu_apriltag/gpu_apriltag.h
@@ -6,21 +6,21 @@
 
 namespace frc971_gpu_apriltag
 {
-    class FRC971GpuApriltagDetectorImpl;
+class FRC971GpuApriltagDetectorImpl;
 
-    class FRC971GpuApriltagDetector
-    {
-    public:
-        FRC971GpuApriltagDetector(const sensor_msgs::CameraInfo::ConstPtr &camera_info);
-        virtual ~FRC971GpuApriltagDetector();
-        void Detect(std::vector<GpuApriltagResult> &results,
-                    std::vector<std::array<cv::Point2f, 4>> &rejected_margin_corners,
-                    std::vector<std::array<cv::Point2f, 4>> &rejected_noconverge_corners,
-                    const cv::Mat &color_image);
+class FRC971GpuApriltagDetector
+{
+public:
+    FRC971GpuApriltagDetector(const sensor_msgs::CameraInfo::ConstPtr &camera_info);
+    virtual ~FRC971GpuApriltagDetector();
+    void Detect(std::vector<GpuApriltagResult> &results,
+                std::vector<std::array<cv::Point2d, 4>> &rejected_margin_corners,
+                std::vector<std::array<cv::Point2d, 4>> &rejected_noconverge_corners,
+                const cv::Mat &color_image);
 
-    private:
-        std::unique_ptr<FRC971GpuApriltagDetectorImpl> impl_;
-    };
+private:
+    std::unique_ptr<FRC971GpuApriltagDetectorImpl> impl_;
+};
 }
 
 #endif
