+++ zebROS_ws/src/gpu_apriltag/src/gpu_apriltag.cu
@@ -9,8 +9,8 @@ namespace frc971_gpu_apriltag
     }
     FRC971GpuApriltagDetector::~FRC971GpuApriltagDetector() = default;
     void FRC971GpuApriltagDetector::Detect(std::vector<GpuApriltagResult> &results,
-                                           std::vector<std::array<cv::Point2f, 4>> &rejected_margin_corners,
-                                           std::vector<std::array<cv::Point2f, 4>> &rejected_noconverge_corners,
+                                           std::vector<std::array<cv::Point2d, 4>> &rejected_margin_corners,
+                                           std::vector<std::array<cv::Point2d, 4>> &rejected_noconverge_corners,
                                            const cv::Mat &color_image)
     {
         impl_->Detect(results, rejected_margin_corners, rejected_noconverge_corners, color_image);
